Added county data to the api through the New York Times dataset
It could be useful to have county data within the United States for analytics and such. The New York Times recently published their dataset on Covid-19, which includes county data. They said it should be updated regularly. The api call is not formatted like many others because of the way the data is processed; this could easily be changed.