Initial changes to run unit tests in docker.
Made some quality of life changes in Dockerfile so that it can reuse installed packages even if you make a code change.
Added .dockerignore to ensure that host machine node_modules do not squash installed node_modules inside the container.
Added `docker-test` command to package.json.
Added `docker-compose.test.yml` to change the command run in the `covid-api-app` container.

Once you've set up your local environment to successfully run the app in docker, you should be able to just run `npm run docker-test`.

It will build the container on every invocation, but docker caching should make the rebuild quick. Note that you can read the exit code of the command, and it will be 0 if all of the tests passed or 1 if there were any failures.